requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
colorama==0.4.3
Django==3.0.7
django-tempus-dominus==5.1.2.13
django-userforeignkey==0.3.0
djangorestframework==3.11.0
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
